Streamline api for cache
Add a type-id for knowledge to put them into the cache.

+++ dwarfhustle-model-db/src/main/java/com/anrisoftware/dwarfhustle/model/db/api/MapChunksStorage.java
@@ -45,7 +45,7 @@ public interface MapChunksStorage extends AutoCloseable {
     /**
      * Returns the {@link MapChunk} with the chunk ID.
      */
-    MapChunk getChunk(int cid);
+    MapChunk getChunk(long cid);
 
     /**
      * Retrieves all {@link MapChunk} chunks.
